Multiple fixes
Fixed the interference of functions with the display in clear_pin function -> caused the pin number to flicker.

Removed unnecessary lines.

First time wizard now starts only on first time booting up.

Added a QR code on a page if there is a problem with the sd card.

Added `sleep_ms(150);` at places where confirming the buttons would couse it to skip a page.

Fixed the position of `"Press Y to continue"` in the `wizard_start()` function to not be rendered outside of a screen.

Imported the qrcode_graphics library.