Bug fixes, changes to visuals
Changed how the entire screen looks. Should now be fullscreen with colorful descriptions next to X and Y buttons.

Added a condition if the directory doesnt exist while opening this page to create the directory.

swapped the logic for a and b buttons to be visually correct.

Added a "No files to delete" text on screen to represent a situation where no file is present.